More sanity refactor
